backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.10.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
eventlet==0.33.3